+++ etils/epy/__init__.py
@@ -19,6 +19,7 @@
 import sys
 
 from etils.epy import _internal
+from etils.epy import typing
 from etils.epy.adhoc_utils.binary_import import binary_adhoc
 from etils.epy.contextlib import ContextManager
 from etils.epy.env_utils import is_notebook