+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Changelog follow https://keepachangelog.com/ format.
     *   Add `epy.is_test` to check whether we're running in a test environment.
     *   Add `epy.typing.Json`.
     *   Add `epy.ExitStack` which allows setting the contextmanagers during init.
+    *   Add proto support for `epy.binary_adhoc`
 *   `exm`:
     *   Add `exm.url_to_python_only_logs()` and `exm.curr_job_name()` to add
         artifact to Python only logs (without all the verbose C++ junk ).