+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Changelog follow https://keepachangelog.com/ format.
 
 ## [Unreleased]
 
+## [1.9.3] - 2024-08-30
+
+*   `eapp`:
+    *   Allow extra kwargs in `eapp.make_flags_parser()`
 *   `epath`:
     *   Fix epath.Path pydantic deserialization for URI-style paths
 *   `epy`:
@@ -436,7 +440,9 @@ Changelog follow https://keepachangelog.com/ format.
 
 *   Automated github release
 
-[Unreleased]: https://github.com/google/etils/compare/v1.9.1...HEAD
+[Unreleased]: https://github.com/google/etils/compare/v1.9.3...HEAD
+[1.9.3]: https://github.com/google/etils/compare/v1.9.2...v1.9.3
+[1.9.2]: https://github.com/google/etils/compare/v1.9.1...v1.9.2
 [1.9.1]: https://github.com/google/etils/compare/v1.9.0...v1.9.1
 [1.9.0]: https://github.com/google/etils/compare/v1.8.0...v1.9.0
 [1.8.0]: https://github.com/google/etils/compare/v1.7.0...v1.8.0